Ensure lspci(1) is installed before running Puppet
Facter rely on lspci(1) to check if the system is a physical or a
virtual one.  When lspci(1) is not found, facter thinks the node is a
physical one, which has an incidence on provisionning (e.g. ensuring
ntp(8) is installed, configured and running which makes basically no
sense on a virtual machine).